+++ packages/deploy-config-sub-generator/test/headless/app-headless.test.ts
@@ -48,6 +48,10 @@ jest.mock('@sap/mta-lib', () => {
     };
 });
 
+jest.mock('os-name', () => {
+    return () => 'mocked-os';
+});
+
 jest.mock('child_process');
 let spawnMock: jest.SpyInstance;